+++ CHANGELOG.md
@@ -9,6 +9,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - Add `PriceRanges` property to `searchQuery`.
 
+## [1.7.2] - 2018-08-03
+### Fixed
+- `Data Layer` to be recreated after each page change.
+
+## [1.7.1] - 2018-08-02
+### Fixed
+- Information pushed to the dataLayer on `ProductSearchContextProvider`
+
 ## [1.7.0] - 2018-08-01
 ### Added
 - Add `<title>` and `<meta>` in store. These fields are available on admin settings.